Splits store into multiple files
