+++ src/constants/index.js
@@ -1,0 +1,1 @@
+export const BASE_URL = "https://api.noroff.dev/api/v1/holidaze";
